requirements.txt
# Веб-фреймворк и сервер
fastapi==0.109.0
uvicorn[standard]==0.27.0

# ORM для работы с базой (выбираем самую популярную — SQLAlchemy)
sqlalchemy==2.0.25

# Драйвер для PostgreSQL (бинарная версия проще ставится в Docker)
asyncpg==0.29.0

# Миграция БД
alembic==1.13.1

# Работа с переменными окружения (Pydantic settings)
pydantic-settings==2.1.0
